+++ Engine/Engine/Models/Block.cs
@@ -6,7 +6,7 @@ namespace Engine.Engine.models
     public class Block : SpriteOverlay
     {
         // Token: 0x06000002 RID: 2 RVA: 0x00002078 File Offset: 0x00000278
-        public Block(double x, double y, int Id, int size, IDrawer engine,IIdProcessor processor) : base(new Sprite(), x, y, Id, engine)
+        public Block(int x, int y, int Id, int size, IDrawer engine,IIdProcessor processor) : base(new Sprite(), x, y, Id, engine)
         {
             processor.ProcessSprite(this);
             Sprite.image = 63;
